Updated AA Sheet
Starting at B86 column index is the new improved AA coefficient calculator. Previously it is assumed that roughly 50% of the total AA damage will be used to shoot down planes. That was a very rough estimate. Now implemented logic that gives you the expected mean coefficient, which would produce the expected mean damage for a given # of planes. Ranges from 1-infinity.